add categorys handle the stock when a bill is submitted (javanullpointerexception to resolve but the stock is correctly updated in the debug)

+++ ProjAppliWeb/src/main/java/m2/project/controller/CategoryController.java
@@ -8,13 +8,11 @@ import javax.validation.Valid;
 import m2.project.model.Category;
 import m2.project.model.ErrorMessage;
 import m2.project.model.JsonResponse;
-
-
+import m2.project.model.TVA;
+import m2.project.repository.TVARepository;
 import m2.project.service.CategoryService;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
-
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -31,6 +29,8 @@ public class CategoryController {
 
 	@Autowired
 	private CategoryService categoryService;
+	@Autowired
+	private TVARepository TVARepository;
 	
 	@RequestMapping(value = "/category", method = RequestMethod.GET)
 	public String categorysList(Model model,@ModelAttribute Category category){//,Pageable pageable,@ModelAttribute Category cat) {
@@ -50,18 +50,25 @@ public class CategoryController {
 	// Submit create / edit  form - Ajax 
 	@RequestMapping(value = "/category", method = RequestMethod.POST, produces={"application/json"})
 	@ResponseBody
-	public JsonResponse ajaxSubmitCategoryForm(Model model, @ModelAttribute(value = "category") @Valid Category category, BindingResult result) {
+	public JsonResponse ajaxSubmitCategoryForm(Model model,@RequestParam(value="tvaid")long tvaid, @RequestParam(value="tva")double tva,@ModelAttribute(value = "category") @Valid Category category, BindingResult result) {
 		JsonResponse res = new JsonResponse();
 		if (!result.hasErrors()) {
 		
 
-			if(!(categoryService.findOne(category.name).isEmpty()))
-				categoryService.save(category);
-			res.setStatus("SUCCESS");
+			//if(!(categoryService.findOne(category.name).isEmpty()))
+				//categoryService.save(category);
+			//res.setStatus("SUCCESS");
 
+				
+			
 			
 			try {
+				TVA newTVA = new TVA(tva);
+				TVARepository.save(newTVA);
+					
+				category.setTVA(newTVA);
 				categoryService.save(category);
+
 				res.setStatus("SUCCESS");
 			}
 			catch(Exception e) {
@@ -73,7 +80,7 @@ public class CategoryController {
 			List<FieldError> allErrors = result.getFieldErrors();
 			List<ErrorMessage> errorMesages = new ArrayList<ErrorMessage>();
 			for (FieldError objectError : allErrors) {
-				errorMesages.add(new ErrorMessage(objectError.getField(), objectError.getDefaultMessage()));
+			errorMesages.add(new ErrorMessage(objectError.getField(), objectError.getDefaultMessage()));
 			}
 			res.setErrorMessageList(errorMesages);
 		}
@@ -87,11 +94,24 @@ public class CategoryController {
 		return categoryService.findOne(id);
 	}
 	
-	@RequestMapping(value = "/category/delete", method = RequestMethod.GET)
-	public String deleteCategory(@RequestParam("id") Long id, Model model) {
-		categoryService.delete(id);
-		return "redirect:/category";
+	@RequestMapping(value = "/category/delete", method = RequestMethod.GET,produces={"application/json"})
+	public @ResponseBody JsonResponse deleteCategory(@RequestParam("id") Long id, Model model) {
+		
+		JsonResponse res = new JsonResponse();
+		
+
+			if((categoryService.findOne(id).getProducts()==null||categoryService.findOne(id).getProducts().isEmpty())){
+				categoryService.delete(id);
+			res.setStatus("OK");
+			}else{
+				res.setStatus("PASOK");
+			}
+			
+		
+	
+		
+		return res;
+	
 	}
-	
-	
+		
 }
